Добавляет MovieInfo
