+++ src/components/movie-page/movie-page.jsx
@@ -6,11 +6,11 @@ import PageHeader from '../page-header/page-header.jsx';
 import UserMenu from '../user-menu/user-menu.jsx';
 import PageContent from '../page-content/page-content.jsx';
 import PageFooter from '../page-footer/page-footer.jsx';
+import MovieInfo from '../movie-info/movie-info.jsx';
 
 
 const MoviePage = ({activeMovie, movies, onMovieTitleClick, onMovieCardMouseEnter}) => {
-  const {id, title, genre, releaseDate, image, ratingScore, ratingLevel, ratingCount, text, director, starring} = activeMovie;
-  const shownActors = starring.slice(0, 4).join(`, `);
+  const {id, title, genre, releaseDate, image} = activeMovie;
   return (
     <Fragment>
       <section key={id} className="movie-card movie-card--full">
@@ -18,13 +18,10 @@ const MoviePage = ({activeMovie, movies, onMovieTitleClick, onMovieCardMouseEnte
           <div className="movie-card__bg">
             <img src={`img/${image}.jpg`} alt={title} />
           </div>
-
           <h1 className="visually-hidden">WTW</h1>
-
           <PageHeader />
 
           <div className="movie-card__wrap">
-
             <div className="movie-card__desc">
               <h2 className="movie-card__title">{title}</h2>
               <p className="movie-card__meta">
@@ -36,49 +33,12 @@ const MoviePage = ({activeMovie, movies, onMovieTitleClick, onMovieCardMouseEnte
                 isLogged={true}
               />
             </div>
-
           </div>
         </div>
 
-        <div className="movie-card__wrap movie-card__translate-top">
-          <div className="movie-card__info">
-            <div className="movie-card__poster movie-card__poster--big">
-              <img src={`img/${image}.jpg`} alt={title} width="218" height="327" />
-            </div>
-
-            <div className="movie-card__desc">
-              <nav className="movie-nav movie-card__nav">
-                <ul className="movie-nav__list">
-                  <li className="movie-nav__item movie-nav__item--active">
-                    <a href="#" className="movie-nav__link">Overview</a>
-                  </li>
-                  <li className="movie-nav__item">
-                    <a href="#" className="movie-nav__link">Details</a>
-                  </li>
-                  <li className="movie-nav__item">
-                    <a href="#" className="movie-nav__link">Reviews</a>
-                  </li>
-                </ul>
-              </nav>
-
-              <div className="movie-rating">
-                <div className="movie-rating__score">{ratingScore}</div>
-                <p className="movie-rating__meta">
-                  <span className="movie-rating__level">{ratingLevel}</span>
-                  <span className="movie-rating__count">{ratingCount} ratings</span>
-                </p>
-              </div>
-
-              <div className="movie-card__text">
-                {text.join(` `)}
-                <p
-                  className="movie-card__director"><strong>Director: {director}</strong></p>
-
-                <p className="movie-card__starring"><strong>Starring: {shownActors} and other</strong></p>
-              </div>
-            </div>
-          </div>
-        </div>
+        <MovieInfo
+          activeMovie={activeMovie}
+        />
       </section>
 
       <PageContent>
