+++ src/components/page-footer/page-footer.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 const PageFooter = () => (
   <footer className="page-footer">
     <div className="logo">
-      <a className="logo__link logo__link--light">
+      <a className="logo__link logo__link--light" href="main.html">
         <span className="logo__letter logo__letter--1">W</span>
         <span className="logo__letter logo__letter--2">T</span>
         <span className="logo__letter logo__letter--3">W</span>
